+++ NEWS.md
@@ -3,9 +3,18 @@
 
 # antaresEditObject 0.9.4.9000
 
+NEW FEATURES :
+
+* `createDistrict()` uses a specific endpoint to create the district
+* `editDistrict()` allows the user to edit a district. Use of a specific endpoint in API mode.
+
 BUGFIXES :    
 
 * `.get_version_solver_from_path_solver()` get solver version by running command line instead of searching in the path.
+
+OPTIMIZATION : 
+
+* `createBindingConstraintBulk()` reads links and clusters of the study once
 
 
 # antaresEditObject 0.9.3
@@ -21,6 +30,7 @@ NEW FEATURES :
 * `createCluster()` :   
   - **Dynamic groups** (The group parameter is now dynamic and has no restrictions. The default group value for Renewable depends on the version: Other RES 1 if < 930, Other if >= 930).
 * `updateScenarioBuilder()` New types of series "sts" and "sta" ("sct apports level" and "sct contraintes") is available
+
 
 BUGFIXES :